Updated README.md to capture the future work
Future work includes a lot of ideas around how to make improvements to the current project.

There are a few ideas which are not even worth discussing further and those will be removed soon. These ideas need to go in as issues. I will be taking care of that in the next few days.